Move language and region selection to script and add to settings dialog
