+++ resources/settings.xml
@@ -857,7 +857,17 @@
                     <control format="string" type="spinner"/>
                 </setting>
             </group>
-            <group id="5" label="30527">
+            <group id="5" label="14222">
+                <setting id="youtube.language_region.configure" type="action" label="30527" help="">
+                    <level>0</level>
+                    <constraints>
+                        <allowempty>true</allowempty>
+                    </constraints>
+                    <data>RunScript($ID,config/language_region)</data>
+                    <control format="action" type="button">
+                        <close>true</close>
+                    </control>
+                </setting>
                 <setting id="youtube.language" type="string" label="30523" help="">
                     <level>0</level>
                     <default>en-US</default>
